+++ src/collective/cover/layout.py
@@ -68,9 +68,9 @@ class PageLayout(grok.View):
             if section['type'] == u'tile':
                 tile_url = '@@{0}/{1}'.format(section.get('tile-type'), section.get('id'))
                 tile_conf = self.context.restrictedTraverse(tile_url.encode()).get_tile_configuration()
-                if tile_conf.get('css_class'):
-                    css_class = tile_conf.get('css_class', '')
-                    section['class'] = '{0} {1}'.format(section.get('class'), css_class)
+                css_class = tile_conf.get('css_class', '')
+                section['class'] = '{0} {1}'.format(section.get('class'), css_class)
+
                 return self.tile(section=section, mode=mode, tile_url=tile_url)
         else:
             return self.generalmarkup(section=section, mode=mode)