+++ src/collective/cover/tests/test_upgrades.py
@@ -5,6 +5,10 @@ from collective.cover.upgrades import register_available_tiles_record
 from collective.cover.upgrades import register_styles_record
 from collective.cover.upgrades import rename_content_chooser_resources
 from collective.cover.upgrades import issue_244
+from collective.cover.upgrades import update_styles_record_4_5
+from collective.cover.upgrades import set_new_default_class_4_5
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
 from plone.registry.interfaces import IRecordAddedEvent
 from plone.registry.interfaces import IRegistry
 from zope.component import eventtesting
@@ -116,6 +120,40 @@ class Upgrade4to5TestCase(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.portal.invokeFactory('Folder', 'test-folder')
+        setRoles(self.portal, TEST_USER_ID, ['Member'])
+        self.folder = self.portal['test-folder']
+        self.folder.invokeFactory('collective.cover.content', 'cover',
+                                  template_layout='Layout B')
+        self.cover = self.folder['cover']
+        self.layout_view = self.cover.restrictedTraverse('layout')
+
+        # get tiles and assign css_class as before version 5
+        layout = self.layout_view.get_layout('view')
+        tile_def = layout[0]['children'][0]['children'][0]
+        self.tile1 = self.cover.restrictedTraverse('{0}/{1}'.format(tile_def['tile-type'], tile_def['id']))
+        tile_config = self.tile1.get_tile_configuration()
+        tile_config['css_class'] = {'order': u'0', 'visibility': u'on'}
+        self.tile1.set_tile_configuration(tile_config)
+
+        tile_def = layout[0]['children'][1]['children'][0]
+        self.tile2 = self.cover.restrictedTraverse('{0}/{1}'.format(tile_def['tile-type'], tile_def['id']))
+        tile_config = self.tile2.get_tile_configuration()
+        tile_config['css_class'] = u"--NOVALUE--"
+        self.tile2.set_tile_configuration(tile_config)
+
+        tile_def = layout[1]['children'][0]['children'][0]
+        self.tile3 = self.cover.restrictedTraverse('{0}/{1}'.format(tile_def['tile-type'], tile_def['id']))
+        tile_config = self.tile3.get_tile_configuration()
+        tile_config['css_class'] = u""
+        self.tile3.set_tile_configuration(tile_config)
+
+        tile_def = layout[1]['children'][1]['children'][0]
+        self.tile4 = self.cover.restrictedTraverse('{0}/{1}'.format(tile_def['tile-type'], tile_def['id']))
+        tile_config = self.tile4.get_tile_configuration()
+        tile_config['css_class'] = u"tile-shadow"
+        self.tile4.set_tile_configuration(tile_config)
 
     def test_issue_244(self):
         css_tool = self.portal['portal_css']
@@ -125,3 +163,27 @@ class Upgrade4to5TestCase(unittest.TestCase):
         self.assertNotIn(id, css_tool.getResourceIds())
         issue_244(self.portal)
         self.assertIn(id, css_tool.getResourceIds())
+
+    def test_issue_262(self):
+        registry = getUtility(IRegistry)
+        record = 'collective.cover.controlpanel.ICoverSettings.styles'
+        default_style = u"-Default-|tile-default"
+
+        # default installation includes the '-Default-|tile-default' style
+        self.assertIn(default_style, registry[record])
+
+        # old installations (upgraded up to version 4) didn't include default style
+        register_styles_record(self.portal)
+        self.assertNotIn(default_style, registry[record])
+
+        # upgraded installations (up to version 5) include default style
+        update_styles_record_4_5(self.portal)
+        self.assertIn(default_style, registry[record])
+
+        # after upgrade step, old tiles have a new default value for
+        # css_class field (if they didn't have one)
+        set_new_default_class_4_5(self.portal)
+        self.assertEqual(self.tile1.get_tile_configuration()['css_class'], u"tile-default")
+        self.assertEqual(self.tile2.get_tile_configuration()['css_class'], u"tile-default")
+        self.assertEqual(self.tile3.get_tile_configuration()['css_class'], u"tile-default")
+        self.assertEqual(self.tile4.get_tile_configuration()['css_class'], u"tile-shadow")