+++ src/collective/cover/tiles/configuration.py
@@ -99,6 +99,9 @@ class TilesConfigurationScreen(object):
             order = unicode(fields.index(name))
             # default configuration attributes for all fields
             defaults[name] = {'order': order, 'visibility': u'on'}
+            if name == 'css_class':
+                # css_class, set default
+                defaults[name] = field.default
             if ITextLine.providedBy(field):
                 # field is TextLine, we should add 'htmltag'
                 defaults[name]['htmltag'] = u'h2'